# Backend-Code/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Commerce API Docs",
        "description": "Commerce Application Programme Interface",
        "termsOfService": "This API is subject to terms of services and maybe be changed without service notice",
        "contact": {
            "email": "sjo065@uowmail.edu.au"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": ""
        }
    },
    "servers": [
        {
            "url": "https://http://167.172.62.67/api/v1",
            "description": "Commerce Staging Server"
        },
        {
            "url": "http://localhost/api/v1"
        }
    ],
    "paths": {
...
        "/carts": {
            "get": {
                "tags": [
                    "Carts"
                ],
                "summary": "The Shopping-Cart Item resource",
                "description": "The Shopping-Cart Item resource",
                "operationId": "Api/CartController::index",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "guest user Identifier key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The resource collection",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CartItemResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "post": {
                "tags": [
                    "Carts"
                ],
                "summary": "The request resource.",
                "description": "The request resource.",
                "operationId": "Api/CartController::store",
                "requestBody": {
                    "description": "The resource body content",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CartRequest"
                            },
                            "example": {
                                "product": 2,
                                "key": "3328uqr984yhoflk39e8yQQQQQQUTYT"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Added to cart successfully.",
                                    "status": true,
                                    "data": {
                                        "id": 1,
                                        "name": "lorem ipsum dolum ima ikegwuru"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/carts/{product}": {
            "get": {
                "tags": [
                    "Carts"
                ],
                "summary": "The CartItemResource.",
                "description": "The CartItemResource.",
                "operationId": "Api/CartController::show",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "The Product resource Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The CartItemResource.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CartItemResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Carts"
                ],
                "summary": "The CartItemResource.",
                "description": "The CartItemResource.",
                "operationId": "Api/CartController::destroy",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "The Product resource Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The resource body content",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CartRequest"
                            },
                            "example": {
                                "key": "3328uqr984yhoflk39e8yQQQQQQUTYT"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The CartItemResource.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CartItemResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Carts"
                ],
                "summary": "The CartItemResource.",
                "description": "The CartItemResource.",
                "operationId": "Api/CartController::update",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "The Product resource Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The CartItemResource.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CartItemResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "default": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CartRequest": {
                "properties": {
                    "course_id": {
                        "description": "The Course resource Id",
                        "type": "integer"
                    },
                    "key": {
                        "description": "The Session Identifier UUid.",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "CartItemResource": {
                "properties": {
                    "id": {
                        "description": "The resource Id.",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The resource name.",
                        "type": "string"
                    },
                    "image": {
                        "description": "The resource image.",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "The resource created date.",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "400": {
                "description": "Bad Request.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Bad Request.",
                            "errors": []
                        }
                    }
                }
            },
...
            "403": {
                "description": "Forbidden.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Forbidden.",
                            "errors": []
                        }
                    }
                }
            },
            "404": {
                "description": "Resource Not Found.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Resource Not Found.",
                            "errors": []
                        }
                    }
                }
            },
...
            "422": {
                "description": "Unprocessable Entity.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Unprocessable Entity.",
                            "errors": []
                        }
                    }
                }
            },
...
            "500": {
                "description": "Server Error.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "Server Error.",
                            "errors": []
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/api/v1/login",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}